SUMMARY: After many long hours debugging the scripts trying to determine the cause of the VirtualBox issues, the issues appear to be a combination of libvirt regressions/VirtualBox incompatablities, and other genuine issues related to libvirt/VirtualBox. The only VirtualBox issue that is consistently reproducible is caused by path of the extracted CernVM image file being too long, aside from that the other issues are difficult to consistently reproduce and resolve.
FIX: Fixed an issue where ssh_generate_key would fail if a key was already generated (on some systems), a key pair is now only generated if one does not exist already

FIX: Fixed an issue where in some cases the extracted image would not be found, now the extract virtual image file is always found

FIX: Fixed some scope issues which were persistent when using trace, the issue is still not entirely resolved due to the limitations of bash. A function cannot feasibly return an array, so the arrays HEADER and TEMPLATE_HEADER in particular must be defined outside the scope of the functions which causes issues when used with "call". No web-interface functions will work if used in conjunction with trace/call due to this scope issue

REMOVE: Removed support for generating template header outside of the web-interface as there are scope issues


git-svn-id: http://cernvm-release-testing.googlecode.com/svn@192 e1b191a7-760c-b37a-3d11-b84f7cd3b9c7